E2: Workaround grammar ambiguity
The current grammar is ambiguous because it can recognize a boolean
value in a declaraction list as a boolean value or as a constant id.

This patch is a workaround. Using `/` instead of '|' gives priority
to the left option. This way the parse tree returned will be the desired
one.